tries fixing

+++ project/sound_class.py
@@ -14,50 +14,52 @@ class OnlyAudio:
         self.WIDTH = 2
         self.p = pyaudio.PyAudio()
         self.flag = False
+        self.second = False
 
     def Send_Sound(self,s):
-        self.stream1 = self.p.open(format=self.FORMAT,
-                                  channels=self.CHANNELS,
-                                  rate=self.RATE,
-                                  input=True,
-                                  frames_per_buffer=self.CHUNK)
-        while True:
-            self.Exit()
-            frames = []
-            for i in range(0, int(self.RATE / self.CHUNK * self.RECORD_SECONDS)):
-                try:
+        try:
+            self.stream1 = self.p.open(format=self.FORMAT,
+                                      channels=self.CHANNELS,
+                                      rate=self.RATE,
+                                      input=True,
+                                      frames_per_buffer=self.CHUNK)
+            while True:
+                frames = []
+                for i in range(0, int(self.RATE / self.CHUNK * self.RECORD_SECONDS)):
+                    self.Exit()
                     data = self.stream1.read(self.CHUNK)
                     frames.append(data)
                     s.sendall(data)
-                except:
-                    self.flag = True
-                    break
+        except:
+            self.flag = True
+            self.Exit()
+
 
     def Exit(self):
-        if self.flag:
+        if self.flag and not self.second:
             self.stream1.stop_stream()
             self.stream1.close()
             self.stream2.stop_stream()
             self.stream2.close()
             self.p.terminate()
-            thread.exit()
+            self.second = True
 
     def Get_Sound(self,s):
-        self.stream2 = self.p.open(format=self.p.get_format_from_width(self.WIDTH),
-                                  channels=self.CHANNELS,
-                                  rate=self.RATE,
-                                  output=True,
-                                  frames_per_buffer=self.CHUNK)
-        data = s.recv(1024)
-        frames = []
-        while data != '':
-            self.Exit()
-            try:
+        try:
+            self.stream2 = self.p.open(format=self.p.get_format_from_width(self.WIDTH),
+                                      channels=self.CHANNELS,
+                                      rate=self.RATE,
+                                      output=True,
+                                      frames_per_buffer=self.CHUNK)
+            data = s.recv(1024)
+            frames = []
+            while data != '':
+                self.Exit()
                 self.stream2.write(data)
                 data = s.recv(1024)
-            except:
-                self.flag = True
-                break
-            frames.append(data)
+                frames.append(data)
+        except:
+            self.flag = True
+            self.Exit()
         self.flag = True
         self.Exit()
